requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.28
requests==2.31.0
beautifulsoup4==4.12.3
fake-useragent==1.4.0
playwright==1.42.0
google-genai==0.3.0
python-dotenv==1.0.1
pypdf==4.1.0
aiofiles==23.2.1
pypdfium2==4.28.0
paddleocr==2.7.3
paddlepaddle==2.6.1
opencv-python-headless==4.9.0.80
aiohttp==3.9.3
langchain==0.1.0
langchain-community==0.0.28
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
pydantic==2.6.3
langsmith==0.1.0
striprtf==0.0.26
# pywin32 is required for .doc conversion on Windows. Install manually if needed: pip install pywin32
pywin32==306; sys_platform == 'win32'
